mlibc: Tinycc support: gcc -nostdinc -nostdlib.
* mlibc/libc-gcc+tcc.c: New file.
* mlibc/include/dlfcn.h:
* mlibc/include/errno.h:
* mlibc/include/signal.h:
* mlibc/include/sys/mman.h:
* mlibc/include/sys/time.h:  Add tcc declarations.

+++ mlibc/libc-mes.c
@@ -322,7 +322,7 @@ malloc (size_t size)
 
 #if !FULL_MALLOC
 void *
-realloc (void *p, int size)
+realloc (void *p, size_t size)
 {
   brk (g_brk + size);
   return g_brk;
